move BranchPickerWidget to workbench core, sessions wrapper adds telemetry
Extract the BranchPickerWidget into the workbench layer
(workbench/contrib/chat/browser/widget/input/) following the
ModelPickerWidget pattern. The sessions BranchPicker becomes a thin
wrapper that listens to onDidSelectBranch and fires
reportNewChatPickerClosed telemetry.
