+++ src/vs/workbench/contrib/chat/browser/widget/input/branchPickerWidget.ts
@@ -1,0 +1,165 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as dom from '../../../../../../base/browser/dom.js';
+import { Gesture, EventType as TouchEventType } from '../../../../../../base/browser/touch.js';
+import { renderIcon } from '../../../../../../base/browser/ui/iconLabel/iconLabels.js';
+import { Codicon } from '../../../../../../base/common/codicons.js';
+import { Emitter, Event } from '../../../../../../base/common/event.js';
+import { Disposable, DisposableStore } from '../../../../../../base/common/lifecycle.js';
+import { autorun, IObservable } from '../../../../../../base/common/observable.js';
+import { localize } from '../../../../../../nls.js';
+import { IActionWidgetService } from '../../../../../../platform/actionWidget/browser/actionWidget.js';
+import { ActionListItemKind, IActionListDelegate, IActionListItem } from '../../../../../../platform/actionWidget/browser/actionList.js';
+
+const FILTER_THRESHOLD = 10;
+
+/** Minimal contract the BranchPickerWidget needs to render and interact. */
+export interface IBranchPickerModel {
+	readonly branches: IObservable<readonly string[]>;
+	readonly selectedBranch: IObservable<string | undefined>;
+	readonly disabled: IObservable<boolean>;
+	setBranch(name: string): void;
+}
+
+export interface IBranchSelectionEvent {
+	readonly branchBefore: string | undefined;
+	readonly branchAfter: string;
+}
+
+interface IBranchItem {
+	readonly name: string;
+}
+
+/**
+ * A widget for selecting a git branch. Renders a trigger button that opens an
+ * {@link IActionWidgetService} dropdown populated from an {@link IBranchPickerModel}.
+ *
+ * Fires {@link onDidSelectBranch} when the user picks a branch so that consumers
+ * can handle telemetry or other side-effects.
+ */
+export class BranchPickerWidget extends Disposable {
+
+	private readonly _renderDisposables = this._register(new DisposableStore());
+	private _slotElement: HTMLElement | undefined;
+	private _triggerElement: HTMLElement | undefined;
+
+	private readonly _onDidSelectBranch = this._register(new Emitter<IBranchSelectionEvent>());
+	readonly onDidSelectBranch: Event<IBranchSelectionEvent> = this._onDidSelectBranch.event;
+
+	constructor(
+		private readonly _model: IObservable<IBranchPickerModel | undefined>,
+		@IActionWidgetService private readonly actionWidgetService: IActionWidgetService,
+	) {
+		super();
+
+		this._register(autorun(reader => {
+			const model = this._model.read(reader);
+			if (model) {
+				model.branches.read(reader);
+				model.selectedBranch.read(reader);
+				model.disabled.read(reader);
+			}
+			this._updateTriggerLabel();
+		}));
+	}
+
+	render(container: HTMLElement): void {
+		this._renderDisposables.clear();
+
+		const slot = dom.append(container, dom.$('.sessions-chat-picker-slot'));
+		this._slotElement = slot;
+		this._renderDisposables.add({ dispose: () => slot.remove() });
+
+		const trigger = dom.append(slot, dom.$('a.action-label'));
+		trigger.tabIndex = 0;
+		trigger.role = 'button';
+		this._triggerElement = trigger;
+		this._updateTriggerLabel();
+
+		this._renderDisposables.add(Gesture.addTarget(trigger));
+		for (const eventType of [dom.EventType.CLICK, TouchEventType.Tap]) {
+			this._renderDisposables.add(dom.addDisposableListener(trigger, eventType, (e) => {
+				dom.EventHelper.stop(e, true);
+				this.showPicker();
+			}));
+		}
+
+		this._renderDisposables.add(dom.addDisposableListener(trigger, dom.EventType.KEY_DOWN, (e) => {
+			if (e.key === 'Enter' || e.key === ' ') {
+				dom.EventHelper.stop(e, true);
+				this.showPicker();
+			}
+		}));
+	}
+
+	showPicker(): void {
+		const model = this._model.get();
+		const branches = model?.branches.get() ?? [];
+		if (!this._triggerElement || this.actionWidgetService.isVisible || branches.length === 0 || model?.disabled.get()) {
+			return;
+		}
+
+		const selectedBranch = model?.selectedBranch.get();
+		const items: IActionListItem<IBranchItem>[] = branches.map(branch => ({
+			kind: ActionListItemKind.Action,
+			label: branch,
+			group: { title: '', icon: Codicon.gitBranch },
+			item: { name: branch, checked: branch === selectedBranch || undefined },
+		}));
+
+		const triggerElement = this._triggerElement;
+		const delegate: IActionListDelegate<IBranchItem> = {
+			onSelect: (item) => {
+				this.actionWidgetService.hide();
+				this._onDidSelectBranch.fire({
+					branchBefore: selectedBranch,
+					branchAfter: item.name,
+				});
+				model?.setBranch(item.name);
+			},
+			onHide: () => { triggerElement.focus(); },
+		};
+
+		const totalActions = items.filter(i => i.kind === ActionListItemKind.Action).length;
+
+		this.actionWidgetService.show<IBranchItem>(
+			'branchPicker',
+			false,
+			items,
+			delegate,
+			this._triggerElement,
+			undefined,
+			[],
+			{
+				getAriaLabel: (item) => item.label ?? '',
+				getWidgetAriaLabel: () => localize('branchPicker.ariaLabel', "Branch Picker"),
+			},
+			totalActions > FILTER_THRESHOLD ? { showFilter: true, filterPlaceholder: localize('branchPicker.filter', "Filter branches...") } : undefined,
+		);
+	}
+
+	private _updateTriggerLabel(): void {
+		if (!this._triggerElement) {
+			return;
+		}
+		dom.clearNode(this._triggerElement);
+
+		const model = this._model.get();
+		const isDisabled = model?.disabled.get() ?? true;
+		const label = model?.selectedBranch.get() ?? localize('branchPicker.select', "Branch");
+
+		dom.append(this._triggerElement, renderIcon(Codicon.gitBranch));
+		const labelSpan = dom.append(this._triggerElement, dom.$('span.sessions-chat-dropdown-label'));
+		labelSpan.textContent = label;
+		dom.append(this._triggerElement, renderIcon(Codicon.chevronDown));
+
+		this._triggerElement.ariaLabel = localize('branchPicker.triggerAriaLabel', "Pick Branch, {0}", label);
+
+		this._slotElement?.classList.toggle('disabled', isDisabled);
+		this._triggerElement.setAttribute('aria-disabled', String(isDisabled));
+		this._triggerElement.tabIndex = isDisabled ? -1 : 0;
+	}
+}
